requirements.txt
-i https://pypi.org/simple
anyio==4.4.0; python_version >= '3.8'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
arrow==1.3.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
asttokens==2.4.1
attrs==23.2.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
bleach==6.1.0; python_version >= '3.8'
cffi==1.16.0; python_version >= '3.8'
colorama==0.4.6; sys_platform == 'win32'
comm==0.2.2; python_version >= '3.8'
debugpy==1.8.2; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.2.13; python_version >= '3.8'
django-bootstrap-v5==1.0.11; python_version >= '3.6' and python_version < '4.0'
django-extensions==3.2.3; python_version >= '3.6'
djangorestframework==3.15.2; python_version >= '3.8'
drf-writable-nested==0.7.0; python_version >= '3.7'
entrypoints==0.4; python_version >= '3.6'
executing==2.0.1; python_version >= '3.5'
fastjsonschema==2.20.0
fqdn==1.5.1
idna==3.7; python_version >= '3.5'
ipykernel==6.29.4; python_version >= '3.8'
ipython==8.26.0; python_version >= '3.10'
ipython-genutils==0.2.0
ipywidgets==8.1.3; python_version >= '3.7'
isoduration==20.11.0
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
jsonpointer==3.0.0
jsonschema[format-nongpl]==4.22.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
jupyter==1.0.0
jupyter-client==7.4.9; python_version >= '3.7'
jupyter-console==6.6.3; python_version >= '3.7'
jupyter-core==5.7.2; python_version >= '3.8'
jupyter-events==0.10.0; python_version >= '3.8'
jupyter-server==2.14.1; python_version >= '3.8'
jupyter-server-terminals==0.5.3; python_version >= '3.8'
jupyterlab-pygments==0.3.0; python_version >= '3.8'
jupyterlab-widgets==3.0.11; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib-inline==0.1.7; python_version >= '3.8'
mistune==3.0.2; python_version >= '3.7'
nbclassic==1.1.0; python_version >= '3.7'
nbclient==0.10.0; python_full_version >= '3.8.0'
nbconvert==7.16.4; python_version >= '3.8'
nbformat==5.10.4; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
notebook==6.5.6; python_version >= '3.7'
notebook-shim==0.2.4; python_version >= '3.7'
overrides==7.7.0; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pandocfilters==1.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.4; python_version >= '3.6'
platformdirs==4.2.2; python_version >= '3.8'
prometheus-client==0.20.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psutil==6.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pure-eval==0.2.2
pycparser==2.22; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-json-logger==2.0.7; python_version >= '3.6'
pywin32==306; sys_platform == 'win32' and platform_python_implementation != 'PyPy'
pywinpty==2.0.13; os_name == 'nt'
pyyaml==6.0.1; python_version >= '3.6'
pyzmq==24.0.1; python_version >= '3.6'
qtconsole==5.5.2; python_version >= '3.8'
qtpy==2.4.1; python_version >= '3.7'
referencing==0.35.1; python_version >= '3.8'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3986-validator==0.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rpds-py==0.18.1; python_version >= '3.8'
send2trash==1.8.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
stack-data==0.6.3
terminado==0.18.1; python_version >= '3.8'
tinycss2==1.3.0; python_version >= '3.8'
tornado==6.4.1; python_version >= '3.8'
traitlets==5.14.3; python_version >= '3.8'
types-python-dateutil==2.9.0.20240316; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.12'
tzdata==2024.1; sys_platform == 'win32'
uri-template==1.3.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0; python_version >= '3.8'
widgetsnbextension==4.0.11; python_version >= '3.7'
